refactor: builder module renamed to `src`
